Corrected two problems. First involved sorting on an attribute that contains the empty string (or any other string that won't convert to the stated type). In the past this invoked Tcl_Panic() and dump core with prejudice. The Tcl_Panic() calls have been removed and now such things just sort to the "low" side.
Second problem dealt with serializing to SQLite in the case of a conditional
referential attribute.  Previously, any empty string that was representing the
conditionality was being passe to the INSERT statement and this would cause a
foreign key constraint failure in SQLite.  Now things are intelligent enough to
test for the null and explicitly pass NULL to SQLite as appropriate.  The
internal representation of NULL is in the ::ral::nullValue variable which is by
default, the empty string.